Test: Rename ambiguous "common" to "utils"
