+++ test/unit/components/FormProvider.spec.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { Map } from 'immutable';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
-import { FormProvider, Form } from '../../../lib';
 import { Input } from '../../components';
-import { defer, validationRules, validationMessages } from '../../common';
+import { defer, validationRules, validationMessages } from '../../utils';
+import { FormProvider, Form } from '../../../lib';
 
 describe('FormProvider', function () {
   it('Returns mutable props when not specified otherwise', () => {
